back-end/requirements.txt
asgiref==3.2.7
astroid==2.3.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
googletrans==2.4.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
wrapt==1.11.2
django-cors-headers-3.2.1

